requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
black==22.6.0
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
coverage==6.4.2
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.6
django-dynamic-fixture==3.1.2
django-environ==0.9.0
django-filter==22.1
django-oauth-toolkit==2.1.0
djangorestframework==3.13.1
executing==0.8.3
filelock==3.7.1
h11==0.13.0
identify==2.5.1
idna==3.3
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
jwcrypto==1.3.1
Markdown==3.3.7
matplotlib-inline==0.1.3
model-mommy==2.0.0
mypy-extensions==0.4.3
nodeenv==1.7.0
oauthlib==3.2.0
outcome==1.2.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.19.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.10
virtualenv==20.15.1
wcwidth==0.2.5
wrapt==1.14.1
wsproto==1.1.0
